requirements.txt
# chatbot 
langchain-openai
langchain-core
langchain-community
python-dotenv
streamlit

# for api
langserve
fastapi
uvicorn
sse_starlette